+++ CHANGELOG.md
@@ -10,6 +10,15 @@ This project does not follow SemVer, since modules are independent of each other
 ### Added
 - Add new module: `elasticache`, [#35](https://github.com/dbl-works/terraform/pull/35)
 
+## [v2021.10.08] - 2021-10-08
+### Changed
+- Ignore minor version changes for RDS, [#37](https://github.com/dbl-works/terraform/pull/37)
+
+## [v2021.10.04] - 2021-10-04
+### Changed
+- Add documentation for the Outline VPN setup [#31](https://github.com/dbl-works/terraform/pull/31)
+- Add Cloudwatch dashboard to ECS module, [#36](https://github.com/dbl-works/terraform/pull/36)
+
 ## [v2021.08.24] - 2021-08-24
 ### Added
 - Add new module: `vpn`, [#30](https://github.com/dbl-works/terraform/pull/30)